# src/routes/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
...
      "Record_string.unknown_": {
        "properties": {},
        "additionalProperties": {},
        "type": "object",
        "description": "Construct a type with a set of properties K of type T"
      },
...
      "RecordId": {
        "type": "string",
        "example": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e"
      },
...
      "PlaintextMessage": {
        "properties": {
          "@type": {
            "type": "string"
          },
          "@id": {
            "type": "string"
          },
          "~thread": {
            "properties": {
              "pthid": {
                "type": "string"
              },
              "thid": {
                "type": "string"
              }
            },
            "type": "object"
          },
          "messageType": {
            "type": "string"
          }
        },
        "required": [
          "@type",
          "@id",
          "messageType"
        ],
        "type": "object",
        "additionalProperties": {}
      },
...
      "AutoAcceptCredential": {
        "description": "Typing of the state for auto acceptance",
        "enum": [
          "always",
          "contentApproved",
          "never"
        ],
        "type": "string"
      },
...
      "Record_string.never_": {
        "properties": {},
        "type": "object",
        "description": "Construct a type with a set of properties K of type T"
      },
...
      "AnonCredsAcceptRequestFormat": {
        "$ref": "#/components/schemas/Record_string.never_",
        "description": "This defines the module payload for calling CredentialsApi.acceptRequest. No options are available for this\nmethod, so it's an empty object"
      },
      "JsonLdAcceptRequestFormat": {
        "description": "Format for accepting a jsonld credential request. Optionally allows the verification\nmethod to use to sign the credential.",
        "properties": {
          "verificationMethod": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "CredentialFormatPayload_CredentialFormats.acceptRequest_": {
        "properties": {
          "indy": {
            "$ref": "#/components/schemas/AnonCredsAcceptRequestFormat"
          },
          "jsonld": {
            "$ref": "#/components/schemas/JsonLdAcceptRequestFormat"
          },
          "anoncreds": {
            "$ref": "#/components/schemas/AnonCredsAcceptRequestFormat"
          }
        },
        "type": "object",
        "description": "Get the payload for a specific method from a list of CredentialFormat interfaces and a method"
      },
      "AcceptCredentialRequestOptions": {
        "properties": {
          "credentialRecordId": {
            "$ref": "#/components/schemas/RecordId"
          },
          "credentialFormats": {
            "$ref": "#/components/schemas/CredentialFormatPayload_CredentialFormats.acceptRequest_"
          },
          "autoAcceptCredential": {
            "$ref": "#/components/schemas/AutoAcceptCredential"
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "credentialRecordId"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "DidExchangeState": {
        "description": "Connection states as defined in RFC 0023.",
        "enum": [
          "start",
          "invitation-sent",
          "invitation-received",
          "request-sent",
          "request-received",
          "response-sent",
          "response-received",
          "abandoned",
          "completed"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      },
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "credo-controller",
    "version": "2.0.0",
    "description": "Rest endpoint wrapper for using your agent over HTTP",
    "license": {
      "name": "Apache-2.0"
    },
    "contact": {}
  },
  "paths": {
...
    "/didcomm/proofs/{proofRecordId}/accept-request": {
      "post": {
        "operationId": "AcceptRequest",
        "responses": {
          "200": {
            "description": "ProofRecord",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record_string.unknown_"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "metadata": {},
                      "id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
                      "createdAt": "2022-01-01T00:00:00.000Z",
                      "connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
                      "threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
                      "protocolVersion": "v1"
                    }
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Accept a presentation request as prover by sending an accept request message\nto the connection associated with the proof record.",
        "tags": [
          "DIDComm - Proofs"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "proofRecordId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "comment": {
                    "type": "string"
                  },
                  "filterByNonRevocationRequirements": {
                    "type": "boolean"
                  },
                  "filterByPresentationPreview": {
                    "type": "boolean"
                  }
                },
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/didcomm/credentials/accept-request": {
      "post": {
        "operationId": "AcceptRequest",
        "responses": {
          "200": {
            "description": "CredentialExchangeRecord",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "type": "object"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "_tags": {
                        "state": "offer-sent",
                        "threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
                        "connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b"
                      },
                      "metadata": {
                        "_internal/w3cCredential": {
                          "credentialDefinitionId": "q7ATwTYbQDgiigVijUAej:3:CL:318187:latest",
                          "schemaId": "q7ATwTYbQDgiigVijUAej:2:Employee Badge:1.0"
                        }
                      },
                      "credentials": [],
                      "id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
                      "createdAt": "2022-01-01T00:00:00.000Z",
                      "connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
                      "threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
                      "credentialAttributes": [],
                      "protocolVersion": "v1"
                    }
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Accept a credential request as issuer by sending an accept request message\nto the connection associated with the credential exchange record.",
        "tags": [
          "DIDComm - Credentials"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptCredentialRequestOptions"
              }
            }
          }
        }
      }
    },
...
    "/didcomm/connections": {
      "get": {
        "operationId": "GetAllConnections",
        "responses": {
          "200": {
            "description": "ConnectionRecord[]",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Record_string.unknown_"
                  },
                  "type": "array"
                },
                "examples": {
                  "Example 1": {
                    "value": [
                      {
                        "_tags": {
                          "invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
                          "did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
                          "invitationKey": "9HG4rJFpLiWf56MWxHj9rgdpErFzim2zEpHuxy1dw7oz",
                          "outOfBandId": "edbc89fe-785f-4774-a288-46012486881d",
                          "verkey": "9HG4rJFpLiWf56MWxHj9rgdpErFzim2zEpHuxy1dw7oz"
                        },
                        "metadata": {},
                        "id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
                        "createdAt": "2022-01-01T00:00:00.000Z",
                        "did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
                        "invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
                        "outOfBandId": "edbc89fe-785f-4774-a288-46012486881d"
                      }
                    ]
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Retrieve all connections records",
        "tags": [
          "DIDComm - Connections"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "outOfBandId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Alias",
            "in": "query",
            "name": "alias",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Connection state",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DidExchangeState"
            }
          },
          {
            "description": "My DID",
            "in": "query",
            "name": "myDid",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Their DID",
            "in": "query",
            "name": "theirDid",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Their label",
            "in": "query",
            "name": "theirLabel",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/didcomm/connections/{connectionId}": {
      "get": {
        "operationId": "GetConnectionById",
        "responses": {
          "200": {
            "description": "ConnectionRecord",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record_string.unknown_"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "_tags": {
                        "invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
                        "did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
                        "invitationKey": "9HG4rJFpLiWf56MWxHj9rgdpErFzim2zEpHuxy1dw7oz",
                        "outOfBandId": "edbc89fe-785f-4774-a288-46012486881d",
                        "verkey": "9HG4rJFpLiWf56MWxHj9rgdpErFzim2zEpHuxy1dw7oz"
                      },
                      "metadata": {},
                      "id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
                      "createdAt": "2022-01-01T00:00:00.000Z",
                      "did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
                      "invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
                      "outOfBandId": "edbc89fe-785f-4774-a288-46012486881d"
                    }
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Retrieve connection record by connection id",
        "tags": [
          "DIDComm - Connections"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [
          {
            "description": "Connection identifier",
            "in": "path",
            "name": "connectionId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RecordId"
            }
          }
        ]
      },
      "delete": {
        "operationId": "DeleteConnection",
        "responses": {
          "204": {
            "description": "No content",
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Deletes a connection record from the connection repository.",
        "tags": [
          "DIDComm - Connections"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [
          {
            "description": "Connection identifier",
            "in": "path",
            "name": "connectionId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RecordId"
            }
          }
        ]
      }
    },
    "/didcomm/connections/{connectionId}/accept-request": {
      "post": {
        "operationId": "AcceptRequest",
        "responses": {
          "200": {
            "description": "ConnectionRecord",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record_string.unknown_"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "_tags": {
                        "invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
                        "did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
                        "invitationKey": "9HG4rJFpLiWf56MWxHj9rgdpErFzim2zEpHuxy1dw7oz",
                        "outOfBandId": "edbc89fe-785f-4774-a288-46012486881d",
                        "verkey": "9HG4rJFpLiWf56MWxHj9rgdpErFzim2zEpHuxy1dw7oz"
                      },
                      "metadata": {},
                      "id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
                      "createdAt": "2022-01-01T00:00:00.000Z",
                      "did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
                      "invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
                      "outOfBandId": "edbc89fe-785f-4774-a288-46012486881d"
                    }
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Accept a connection request as inviter by sending a connection response message\nfor the connection with the specified connection id.\n\nThis is not needed when auto accepting of connection is enabled.",
        "tags": [
          "DIDComm - Connections"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [
          {
            "description": "Connection identifier",
            "in": "path",
            "name": "connectionId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RecordId"
            }
          }
        ]
      }
    },
    "/didcomm/connections/{connectionId}/accept-response": {
      "post": {
        "operationId": "AcceptResponse",
        "responses": {
          "200": {
            "description": "ConnectionRecord",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record_string.unknown_"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "_tags": {
                        "invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
                        "did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
                        "invitationKey": "9HG4rJFpLiWf56MWxHj9rgdpErFzim2zEpHuxy1dw7oz",
                        "outOfBandId": "edbc89fe-785f-4774-a288-46012486881d",
                        "verkey": "9HG4rJFpLiWf56MWxHj9rgdpErFzim2zEpHuxy1dw7oz"
                      },
                      "metadata": {},
                      "id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
                      "createdAt": "2022-01-01T00:00:00.000Z",
                      "did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
                      "invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
                      "outOfBandId": "edbc89fe-785f-4774-a288-46012486881d"
                    }
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Accept a connection response as invitee by sending a trust ping message\nfor the connection with the specified connection id.\n\nThis is not needed when auto accepting of connection is enabled.",
        "tags": [
          "DIDComm - Connections"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [
          {
            "description": "Connection identifier",
            "in": "path",
            "name": "connectionId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RecordId"
            }
          }
        ]
      }
    },
    "/didcomm/url/{invitationId}": {
      "get": {
        "operationId": "GetInvitation",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaintextMessage"
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "DIDComm - Connections"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "invitationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}